work_scripts/python_scripts/command_to_server/requirements.txt
paramiko
pyyaml
tqdm
cryptography